upload/requirements.txt
playwright==1.53.0
requests==2.32.4
openai==1.97.0
dspy-ai==2.6.27
langgraph==0.5.2
mem0ai
python-dotenv==1.1.1
